+++ clients/drcachesim/tracer/instr_counter.h
@@ -1,0 +1,59 @@
+/* **********************************************************
+ * Copyright (c) 2011-2022 Google, Inc.  All rights reserved.
+ * Copyright (c) 2010 Massachusetts Institute of Technology  All rights reserved.
+ * **********************************************************/
+
+/*
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * * Redistributions of source code must retain the above copyright notice,
+ *   this list of conditions and the following disclaimer.
+ *
+ * * Redistributions in binary form must reproduce the above copyright notice,
+ *   this list of conditions and the following disclaimer in the documentation
+ *   and/or other materials provided with the distribution.
+ *
+ * * Neither the name of Google, Inc. nor the names of its contributors may be
+ *   used to endorse or promote products derived from this software without
+ *   specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL VMWARE, INC. OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+ * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
+ * DAMAGE.
+ */
+
+#ifndef _INSTR_COUNTER_
+#define _INSTR_COUNTER_ 1
+
+#include "dr_api.h"
+
+namespace dynamorio::drmemtrace {
+
+dr_emit_flags_t
+event_inscount_bb_analysis(void *drcontext, void *tag, instrlist_t *bb, bool for_trace,
+                           bool translating, void **user_data);
+
+dr_emit_flags_t
+event_inscount_app_instruction(void *drcontext, void *tag, instrlist_t *bb,
+                               instr_t *instr, instr_t *where, bool for_trace,
+                               bool translating, void *orig_analysis_data,
+                               void *user_data);
+
+void
+event_inscount_thread_init(void *drcontext);
+
+void
+event_inscount_init();
+
+} // namespace dynamorio::drmemtrace
+
+#endif /* _INSTR_COUNTER_ */
